`prn` module updates
- Removed unnecessary layer to the `prn` module
- Updated `prn` docs
